Adds base config for `DagFactory`
This changes the way the config initialization works for `DagFactory`.
We now support a `base_config` parameter, in addition to the old
`extra_config` one. The whole config is now initialized before the rest
of the parameters (such as `default_args`). First the `base_config` is
loaded, the the config from the airflow database is merged in, and
finally `extra_config` is merged.

This allows the user to provide default configuration option through
`base_config`, override those defaults with the config from the database
and finally override everything else with `extra_config`