Upgrades build system and dependecy management
This commit upgrades the package build system and dependency declaration
to reflect better the way this library is supposed to be used. For
example, we require *some* version of apache beam to be installed in the
target environment, and not a specific version like if this was an
actual deployment target.